Replaces connections-overview with react component
